Add logging handlers to flask.
For some reason the logging mechanism in flask is being a pain in the
ass.  Need to find out why it's not accepting the handlers I'm adding to
it and why it's not using them to send out messages.

Also, need to find a nice way to take a logging configuration file and
setup all the handlers appropriately (would be nice if I could just
override the root log handler from config).