Factor mapid executable into api module.
The flask module should be callable as a WSGI application for
deployments anyway and it makes more sense to call the module directly
when flask is desired to run.  Especially with the tests that will be
coming.  Thus, the new way to run the application should be `python -m
margarine.api`.