Update mock power set generation and selection.
After a bit of review I've found that the old way to generate the
integration tests was incorrect.  The new method uses the unit's mock
mask to determine what needs to be unmasked to generate a test.  If
there are no masked mocks we were already in a mockless world.