Remove mailing list notifications from travis.
Seems that using google groups for our developer mailing list is not
compatible with travis-ci's notifications.  Trying to send an email to
the google groups list from travis is blocked due to membership being
required.  Also, when trying to add the notifications email for travis
as a member we're alerted to the fact that it too is a google group and
can't be added to a google group.