Cleanup put mthod for blend.user.
Switched this method to single quotes for aesthetics.  Also, re-ordered
sections to make everything more local to what it's influencing.  The
ordering now makes the method read better as well as allows for more
short-circuit processing.

I have determined that crafting of the user method may need to be
refactored into it's own private method but we'll doucle check elsewhere
if it's used enough to warrant this.  Also, the test is incorrect in the
assumption that None values should not be passed to through the message
queue.