Add check for Vagrantfile to generated setUp.
The generated tests' setUp now includes a check for a Vagrantfile
(similar to vagrant's check) which will skip the test if it's not found.

This check is contained in a helper function added in this commit as
well.  find_vagrant_file will search the current directory and its
parents for a Vagrantfile.

Issues: #43