Update information merger in blend.user.put.
The properties in blend.user.put needed to be merged before being passed
across the queue.  I have fixed this issue by using the already present
value (if one exists) as the default value in retrieving the form data.
This solves the test and only a review is necessary.

This should fix #17.