Add datastore attributes to blend role.
For some reason the type of the datastore is set to nil on this node.
Also, the username and password need to be set for the mongo deployment.
The username and password should not be hard coded but that's a change
for a future item (probably when we go to production or staging).