Rework article submission top-half.
The correct information now gets passed through the top-half of the
article submission process (the stubbing process if you will).

Also in this commit is a more robust logging mechanism (although I'm
still thinking that logging the uuid5 of the url is a good idea).

A couple of assumptions needed to be fixed in the test itself which
required a copy not an assignment of the article dict being verifeid
versus being passed.  Once these items were in alignment the test was
driving this case quite correctly (as far as I can see).

This should fix #42 but a review will need to done first.