Update test story for #41.
Factor out the updates and create a copy of the data structure being
tested.  These are the same changes with the same reasoning as #42 and
should be dealt with in a similar fashion.

Thus, this issue is now blocked by the merging of #42 as that already solves
the method under inspection.  This work will have to wait until that
branch is merged before progressing in earnest.