Add monkey patch for issue #23.
Looks like the nosetests section of setup.cfg causes issues and needs to
be bypassed when using setup.py.  This also means until we can find a
way to actually fix the issue we can't use setup.py nosetests (which
would make one command for doing all things).

This should fix #23 and allows us to use setup.py for installs, &c
again.  But not for running tests.