Add build to integration process.
I've changed to nosetests for the test runner as this is more
appropriate than calling a main module or a testing module in our code.
I've also added the installation process to the items to be tested.  If
the software doesn't install, then it should be fixed, immediately.

I also added in python2.6 with the removal of the module call but expect
to remove it shortly due to 2.6 being deprecated and our software not
working.  I'm keeping my eye on pika for when we can start supporting
python3.