Convert parameters to crumbs.
This is a pretty large commit.  It not only switches the old parameter
construct out for the one provided by crumbs but updates the parameter
definitions and rectifies the tests as well.  Thus, all references
should be up to date but further testing will be required.

Unfortunately, this commit also breaks backwards compatibility and
requires any installation mechanisms such as the chef cookbook to also
be updated.

This should close issue #55 if it passes review.