Remove chef directory.
I've reduced the vagrant chef ties to the bare minimum (for now) and
inlined the attributes required to get a working environment started by
vagrant up.

I've also removed the berksfile dependency on the margarine cookbook as
a result and updated the vagrantfile so that we no longer deploy
margarine with chef (much faster this way).