Remove references to pyrax.
As requested in issue #57, pyrax should be removed and the body of
articles should be stored in MongoDB's GridFS.  This commit makes those
changes and ensures that the existing tests continue to run.

This should assist the development environment in vagrant as well.  By
not needing a cloud files account we can isolate an environment and run
completely on a developer workstation if necessary.