Add create article API call.
This only partially satisfies issue #58 but also begins work necessary
for issues #72 and #75 as well.

Using kombu makes the message queue portion of margarine pluggable
without adding any complexity.  It also allows pickling as a
serialization mechanism amongst others.

This only completes the top-half of issue #58 though.  It only
implements the necessary changes for tests to pass and blend having
support for creation of an article (popping relevant information into a
message queue).

Next steps will be adding corresponding updates to spread with unit
tests for that functionality and then adding integration tests for blend
and spread with an unmocked queue service (vagrant up queue).