Add worker framework for spread using kombu.
Lots of changes in this that are largely aesthetic.  The major changes
include adding a kombu framework for spread and the first task handler.

The unit tests for this first task has also been added.  This required a
bit of refactoring the test suite.

Work left to do, includes adding system tests that show an article can
be added and read through the API.