Add sanitize worker.
The addition of this work brought about many changes to the fixtures
being stored (not sure we have them right yet but it's closer).  The
addition of the tests increased the need for sharing of mocks and
generalization of some of that code as well as showed wht is required
for generating consistent etags.

Next steps should be integration and system tests for create and read of
an article through the API.