Factor spread.article into package.
As the complexity of sanitization grows it will be nice to take
advantage of the recursive loading of worker modules and have
sanitization of HTML and articles in its own module.  Thus, the first
step was making this re-organized structure pass tests and lint.

Also, in this commit is a change to vagrant that updates us to the
latest LTS release for Ubuntu.  For some reason the box is a 32-bit
machine but it's already much faster and works well for our purposes.
It's also nice to have a more up to date RabbitMQ (the console is super
pretty).